Implemented Day 8 (#20)

+++ src/aoc/day02.rs
@@ -1,20 +1,20 @@
 struct Game {
-    id: u32,
-    red: Vec<u32>,
-    green: Vec<u32>,
-    blue: Vec<u32>,
+    id: u64,
+    red: Vec<u64>,
+    green: Vec<u64>,
+    blue: Vec<u64>,
 }
 
 impl Game {
     pub fn new(line: &str) -> Option<Self> {
         let (game, cubes) = line.split_once(":")?;
-        let id = game.trim().split_once(" ")?.1.parse::<u32>().ok()?;
+        let id = game.trim().split_once(" ")?.1.parse::<u64>().ok()?;
         let mut game = Game { id, red: vec![], green: vec![], blue: vec![] };
         cubes.split(";").flat_map(|x| x.split(",")).for_each(|chunk| {
             match chunk.trim().split_once(" ") {
-                Some((c, "red")) => game.red.push(c.parse::<u32>().unwrap_or(0)),
-                Some((c, "green")) => game.green.push(c.parse::<u32>().unwrap_or(0)),
-                Some((c, "blue")) => game.blue.push(c.parse::<u32>().unwrap_or(0)),
+                Some((c, "red")) => game.red.push(c.parse::<u64>().unwrap_or(0)),
+                Some((c, "green")) => game.green.push(c.parse::<u64>().unwrap_or(0)),
+                Some((c, "blue")) => game.blue.push(c.parse::<u64>().unwrap_or(0)),
                 _ => {}
             };
         });
@@ -22,14 +22,14 @@ impl Game {
     }
 }
 
-pub fn part1(input: &str) -> u32 {
+pub fn part1(input: &str) -> u64 {
     input.split("\n").filter_map(Game::new)
         .filter(|g| g.red.iter().all(|&x| x <= 12) && g.green.iter().all(|&x| x <= 13) && g.blue.iter().all(|&x| x <= 14))
         .map(|g| g.id)
         .sum()
 }
 
-pub fn part2(input: &str) -> u32 {
+pub fn part2(input: &str) -> u64 {
     input.split("\n").filter_map(Game::new)
         .map(|g| g.red.iter().max().unwrap_or(&0) * g.green.iter().max().unwrap_or(&0) * g.blue.iter().max().unwrap_or(&0))
         .sum()
